Fixed Swift 3.0.2 compilation

+++ Wire-iOS/Sources/UserInterface/Collections/CollectionImageCell.swift
@@ -26,7 +26,7 @@ final public class CollectionImageCell: CollectionCell {
     static var imageCache: ImageCache {
         let cache = ImageCache(name: "CollectionImageCell.imageCache")
         cache.maxConcurrentOperationCount = 4
-        cache.totalCostLimit = 1024 * 1024 * 20 // 20 MB
+        cache.totalCostLimit = UInt(1024 * 1024 * 20) // 20 MB
         cache.qualityOfService = .utility
     
         return cache
